+++ frontend/src/metabase/querying/components/FilterModal/utils/index.ts
@@ -1,0 +1,3 @@
+export * from "./filters";
+export * from "./modal";
+export * from "./search";
