+++ frontend/src/metabase/querying/components/FilterModal/utils/modal.ts
@@ -1,0 +1,13 @@
+import { t } from "ttag";
+import type { GroupItem } from "../types";
+
+export function getModalTitle(groupItems: GroupItem[]) {
+  return groupItems.length === 1
+    ? t`Filter ${groupItems[0].displayName} by`
+    : t`Filter by`;
+}
+
+export function getModalWidth(groupItems: GroupItem[]) {
+  const maxWidth = groupItems.length > 1 ? "70rem" : "55rem";
+  return `min(98vw, ${maxWidth})`;
+}
